+++ testapp/home/models.py
@@ -1,4 +1,6 @@
-from wagtail import VERSION as WAGTAIL_VERSION
+from wagtail.admin.panels import FieldPanel as StreamFieldPanel
+from wagtail.fields import StreamField
+from wagtail.models import Page
 
 from wagtailnhsukfrontend.blocks import (ActionLinkBlock, CardBasicBlock,
                                          CardClickableBlock, CardFeatureBlock,
@@ -9,15 +11,6 @@ from wagtailnhsukfrontend.blocks import (ActionLinkBlock, CardBasicBlock,
                                          InsetTextBlock, SummaryListBlock,
                                          WarningCalloutBlock)
 from wagtailnhsukfrontend.mixins import HeroMixin, ReviewDateMixin
-
-if WAGTAIL_VERSION >= (3, 0):
-    from wagtail.admin.panels import FieldPanel as StreamFieldPanel
-    from wagtail.fields import StreamField
-    from wagtail.models import Page
-else:
-    from wagtail.admin.edit_handlers import StreamFieldPanel
-    from wagtail.core.models import Page
-    from wagtail.core.fields import StreamField
 
 
 class HomePage(HeroMixin, ReviewDateMixin, Page):
@@ -36,20 +29,7 @@ class HomePage(HeroMixin, ReviewDateMixin, Page):
         ('image', ImageBlock()),
         ('warning_callout', WarningCalloutBlock()),
         ('summary_list', SummaryListBlock()),
-    ], use_json_field=True) if WAGTAIL_VERSION >= (3, 0) else StreamField([
-        ('action_link', ActionLinkBlock()),
-        ('care_card', CareCardBlock()),
-        ('details', DetailsBlock()),
-        ('do_list', DoBlock()),
-        ('dont_list', DontBlock()),
-        ('expander', ExpanderBlock()),
-        ('expander_group', ExpanderGroupBlock()),
-        ('feature_card', CardFeatureBlock()),
-        ('inset_text', InsetTextBlock()),
-        ('image', ImageBlock()),
-        ('warning_callout', WarningCalloutBlock()),
-        ('summary_list', SummaryListBlock()),
-    ])
+    ], use_json_field=True)
 
     content_panels = Page.content_panels + HeroMixin.content_panels + [
         StreamFieldPanel('body'),
@@ -75,13 +55,7 @@ class HubsPage(Page):
         ('card_image', CardImageBlock()),
         ('card_feature', CardFeatureBlock()),
         ('card_group', CardGroupBlock()),
-    ], use_json_field=True) if WAGTAIL_VERSION >= (3, 0) else StreamField([
-        ('card_basic', CardBasicBlock()),
-        ('card_clickable', CardClickableBlock()),
-        ('card_image', CardImageBlock()),
-        ('card_feature', CardFeatureBlock()),
-        ('card_group', CardGroupBlock()),
-    ])
+    ], use_json_field=True)
     content_panels = Page.content_panels + [
         StreamFieldPanel('body'),
     ]