+++ testapp/testapp/urls.py
@@ -1,15 +1,9 @@
 from django.conf import settings
-from django.urls import include, path
 from django.contrib import admin
-from wagtail import VERSION as WAGTAIL_VERSION
+from django.urls import include, path
+from wagtail import urls as wagtail_urls
 from wagtail.admin import urls as wagtailadmin_urls
 from wagtail.documents import urls as wagtaildocs_urls
-
-if WAGTAIL_VERSION >= (3, 0):
-    from wagtail import urls as wagtail_urls
-else:
-    from wagtail.core import urls as wagtail_urls
-
 
 urlpatterns = [
     path("django-admin/", admin.site.urls),