+++ tests/test_card_block.py
@@ -1,6 +1,6 @@
+import pytest
 from bs4 import BeautifulSoup
 from django.test import Client
-import pytest
 
 
 @pytest.mark.django_db